requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3=1.26.0
Django==4.1.2
django-cors-headers==3.13.0
django-environ=0.9.0
django-filter==22.1
djangorestframework==3.14.0
dynamorm=0.11.0
marshmallow=3.18.0
pytz==2022.4
sqlparse==0.4.3
